fix put, get by id, made new files for functions
