+++ helpers/routerFn.js
@@ -1,0 +1,83 @@
+const contactService = require('..//models/contacts');
+const contactAddSchema = require('./validateData');
+const HttpError = require('./HttpError');
+
+const listContact = async (req, res, next) => {
+  try {
+    const result = await contactService.listContacts();
+    res.json(result);
+  } catch (error) {
+    next(error);
+  }
+};
+
+const getById = async (req, res, next) => {
+  try {
+    const { contactId } = req.params;
+    const result = await contactService.getContactById(contactId);
+    if (result === null) {
+      throw HttpError(404);
+    }
+    res.json(result);
+  } catch (error) {
+    next(error);
+  }
+};
+
+const addContact = async (req, res, next) => {
+  try {
+    const { error } = contactAddSchema.validate(req.body);
+    if (error) {
+      throw HttpError(400, error.message);
+    }
+    const result = await contactService.addContact(req.body);
+    res.status(201).json(result);
+  } catch (error) {
+    next(error);
+  }
+};
+
+const removeContact = async (req, res, next) => {
+  try {
+    const { contactId } = req.params;
+    const result = await contactService.removeContact(contactId);
+    if (!result) {
+      throw HttpError(404);
+    }
+    res.json({ message: 'contact deleted' });
+  } catch (error) {
+    next(error);
+  }
+};
+
+const updateContact = async (req, res, next) => {
+  try {
+    const body = req.body;
+
+    if (Object.keys(body).length === 0) {
+      throw HttpError(400, 'missing fields');
+    }
+    const { error } = contactAddSchema.validate(req.body);
+    if (error) {
+      throw HttpError(400, error.message);
+    }
+
+    console.log('reqBody', req.body);
+    const { contactId } = req.params;
+    const result = await contactService.updateContactById(contactId, req.body);
+    if (!result) {
+      throw HttpError(404);
+    }
+    res.json(result);
+  } catch (error) {
+    next(error);
+  }
+};
+
+module.exports = {
+  listContact,
+  getById,
+  addContact,
+  removeContact,
+  updateContact,
+};
